requirements.txt
python-twitter
vaderSentiment
peewee
